refactor(inference): split context window module
